feat: 🎸 api client
Add api client with shared types for services to make it possible to
manage the api data model as a typed resource in the client. Add an API
Tests suite to testing using an API Client. Update to general
configuration to support esm and commonjs for types. Update to general
tooling configuration (scripts)